requirement.txt
flask==3.0.3
google-auth==2.29.0